requirements.txt
Flask==3.1.0
redis==4.3.4
pandas==2.2.3
rpi-lgpio==0.6
# libcamera
